Moved Discord bot to separate init, separated dev mode and constants
