+++ config.json.default
@@ -1,0 +1,7 @@
+{
+    "discordApiToken": null,
+    "accessKeyId": null,
+    "secretAccessKey": null,
+    "tradeMnemonic": null,
+    "treasuryMnemonic": null
+}
