+++ constants.js
@@ -1,0 +1,7 @@
+exports.ethereumHost = 'ethereum.exokit.org';
+exports.usersTableName = 'users';
+exports.prefix = '.';
+exports.storageHost = 'https://ipfs.exokit.org';
+exports.previewHost = 'https://preview.exokit.org';
+exports.previewExt = 'png';
+exports.treasurerRoleName = 'Treasurer';
